+++ src/routes/datasets/Datasets.tsx
@@ -2,8 +2,7 @@
 
 /** @jsx jsx */
 import DatasetSider from './components/DatasetSider';
-import { Flex } from '@chakra-ui/react';
-import { Tbody, Td } from '@chakra-ui/table';
+import { Flex, Tbody, Td } from '@chakra-ui/react';
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { css, jsx } from '@emotion/react';