+++ src/routes/tasks/Tasks.tsx
@@ -4,8 +4,7 @@
 import { useState } from 'react';
 
 import TaskSider from './components/TaskSider';
-import { Flex } from '@chakra-ui/react';
-import { Tbody, Td } from '@chakra-ui/table';
+import { Flex, Tbody, Td } from '@chakra-ui/react';
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { css, jsx } from '@emotion/react';
 import styled from '@emotion/styled';